--- data/Duke.txt
@@ -1,3 +1,0 @@
-T | 0 | borrow book
-T | 0 | eat
-E | 0 | meeting|  May 1 2020
