+++ data/duke.txt
@@ -1,0 +1,6 @@
+T | 0 |  read book
+D | 1 |  abc   |   Friday
+E | 1 |  project meeting  |  4pm
+E | 0 |  supper  |  tonight
+T | 1 |  eat
+E | 0 | birthday party | Tuesday night
