--- src/main/java/Duke/Exception/EmptyFindContent.java
@@ -1,8 +1,0 @@
-package Duke.Exception;
-
-public class EmptyFindContent extends Exception {
-    @Override
-    public String getMessage() {
-        return "Find command must follow by a keyword!";
-    }
-}
