+++ src/main/java/Event.java
@@ -1,0 +1,19 @@
+public class Event extends Task{
+    protected String at;
+
+    public Event(String description, String at) {
+        super(description);
+        this.at = at;
+        this.type = "E";
+    }
+
+    @Override
+    public String getTime() {
+        return " | " + at;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (at: " + at + ")";
+    }
+}
