+++ src/main/java/NoSuchCommandException.java
@@ -1,0 +1,7 @@
+public class NoSuchCommandException extends Exception{
+
+    @Override
+    public String getMessage() {
+        return "☹ OOPS!!! I'm sorry, but I don't know what that means :-(";
+    }
+}
