+++ src/main/java/Todo.java
@@ -1,0 +1,11 @@
+public class Todo extends Task{
+    public Todo(String description) {
+        super(description);
+        this.type = "T";
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
