--- src/test/java/Duke/DukeTest.java
@@ -1,12 +1,0 @@
-package Duke;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class DukeTest {
-    @Test
-    public void dummyTest(){
-        assertEquals(2, 2);
-    }
-}
